Began to rename components so that the names are more general and can fit other uses throughout the app
