--- src/components/AlphabetLetter/index.js
@@ -1,1 +1,0 @@
-export { default } from './AlphabetLetter.js'
