+++ src/components/ButtonList/ButtonList.styles.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const List = styled.div`
+export const ButtonMatrix = styled.div`
   width: 360px;
   display: flex;
   flex-direction: row;
@@ -12,7 +12,7 @@ export const List = styled.div`
   padding-bottom: 20px;
 `;
 
-export const LetterRow = styled.div`
+export const ButtonRow = styled.div`
   display: flex;
   justify-content: center;
   flex-direction: row;
