+++ src/components/index.js
@@ -1,8 +1,8 @@
-export { default as AlphabetLetter } from './AlphabetLetter'
-export { default as AlphabetList } from './AlphabetList'
-export { default as LanguageSwitch } from './LanguageSwitch'
-export { default as LetterDetail } from './LetterDetail'
-export { default as LetterTiles } from './LetterTiles'
-export { default as NameButton } from './NameButton'
-export { default as SpellingCard } from '.SpellingCard'
-export { default as TileLetter } from './TileLetter'
+export { default as Button } from "./Button";
+export { default as ButtonList } from "./ButtonList";
+// export { default as LanguageSwitch } from './LanguageSwitch'
+export { default as LetterDetail } from "./LetterDetail";
+export { default as LetterTiles } from "./LetterTiles";
+export { default as NameButton } from "./NameButton";
+// export { default as SpellingCard } from '.SpellingCard'
+export { default as TileLetter } from "./TileLetter";
